% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvcor.R
\name{lvcor}
\alias{lvcor}
\title{Compute rotation of latent variables in fitted formative-reflective models}
\usage{
lvcor(pm)
}
\arguments{
\item{pm}{an object of class \code{pathmod}}
}
\value{
An object of class \code{lvcor} with slots:
\describe{
    \item{\code{alphas}}{a numeric vector with the values of \eqn{\alpha}}
    \item{\code{xicor}}{\eqn{p} x \eqn{p} x \code{length(alphas)} array with the correlation between the exogenous latent
    variables for \eqn{\alpha = 0} and the other values of \eqn{\alpha}}
    \item{\code{etacor}}{\eqn{q} x \eqn{q} x \code{length(alphas)} array with the correlation between the endogenous latent
    variables for \eqn{\alpha = 0} and the other values of \eqn{\alpha}}
}
}
\description{
It computes the correlations (rotation) among latent variables as alpha changes.
}
