% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathmod.R
\name{pathmod}
\alias{pathmod}
\title{Extract latent variables in a formative-reflective model}
\usage{
pathmod(xgroups, ygroups, data = NULL, cov = NULL, alphas = seq(0, 10)/10,
  xmean = NULL, ymean = NULL, xsd = NULL, ysd = NULL)
}
\arguments{
\item{xgroups}{a list containing \eqn{p} char vectors with the names of the variables of each \eqn{x}-group.}

\item{ygroups}{a list containing \eqn{q} char vectors with the names of the variables of each \eqn{y}-group.}

\item{data}{the data frame with the observations (optional, but needed if you want the latent
variable scores).}

\item{cov}{the named covariance matrix of the variables (if you specify the \code{data} argument and
leave it \code{null}, then \code{cov} is assigned the correlation matrix of \code{data}).}

\item{alphas}{a numeric vector with values in the range [0, 1] for the \eqn{\alpha}-weights.}

\item{xmean}{optional parameter to center the \eqn{x}-variables taken from \code{data}, otherwise equal to their sample means}

\item{ymean}{optional parameter to center the \eqn{y}-variables taken from \code{data}, otherwise equal to their sample means}

\item{xsd}{optional parameter to scale the \eqn{x}-variables taken from \code{data}, otherwise equal to their sample standard deviations}

\item{ysd}{optional parameter to scale the \eqn{y}-variables taken from \code{data}, otherwise equal to their sample standard deviations}
}
\value{
An object of class \code{pathmod} with slots:
\describe{
    \item{\code{omegas}}{list of matrices \eqn{\Omega}, one for each element in \code{alpha}}
    \item{\code{gammas}}{list of matrices \eqn{\Gamma}, one for each element in \code{alpha}}
    \item{\code{lambdas}}{list of matrices \eqn{\Lambda}, one for each element in \code{alpha}}
    \item{\code{xlosses}}{vector of \eqn{x}-side losses, one for each element in \code{alpha}}
    \item{\code{ylosses}}{vector of \eqn{y}-side losses, one for each element in \code{alpha}}
    \item{\code{alphas}}{same as input argument (values of \eqn{\alpha})}
    \item{\code{p}}{number of \eqn{x}-groups}
    \item{\code{q}}{number of \eqn{y}-groups}
    \item{\code{Sxx}}{variance-covariance matrix of exogenous variables}
    \item{\code{Syx}}{covariance matrix with covariances between endogenous and exogenous variables}
    \item{\code{Syy}}{variance-covariance matrix of endogenous variables}
    \item{\code{xis}}{three-dimensional array with scores of the exogenous latent variables, one for each element in \code{alpha} (available only if \code{data} is assigned)}
    \item{\code{etas}}{three-dimensional array with scores of the endogenous latent variables, one for each element in \code{alpha} (available only if \code{data} is assigned)}
    \item{\code{xgroups}}{the argument \code{xgroups}}
    \item{\code{ygroups}}{the argument \code{ygroups}}
    \item{\code{xmean}}{the supplied argument \code{xmean} or the means of the \eqn{x}-variables in \code{data}}
    \item{\code{ymean}}{the supplied argument \code{ymean} or the means of the \eqn{y}-variables in \code{data}}
    \item{\code{xsd}}{the supplied argument \code{xsd} or the standard deviations of the \eqn{x}-variables in \code{data}}
    \item{\code{ysd}}{the supplied argument \code{ysd} or the standard deviations of the \eqn{y}-variables in \code{data}}
}
}
\description{
Apply the method of Fattore, Pelagatti, Vittadini (2017) to extract latent variables in a formative-reflective model.
}
\examples{
data(russett)
xg <- list(c("gini", "farm", "rent"), c("gnpr", "labo"))
yg <- list(c("inst", "ecks", "death"))
pm <- pathmod(xg, yg, data = russett)
}
