% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lvcor.R
\name{plot.lvcor}
\alias{plot.lvcor}
\title{Plot method for the rotation of latent variables in a fitted formative-reflective model}
\usage{
\method{plot}{lvcor}(x, lv = c("xi", "eta"), ...)
}
\arguments{
\item{x}{an object of class \code{lvcor}}

\item{lv}{a string: \code{"xi"} to compute exogenous latent variables' rotation, \code{"eta"} to compute
endogenous latent variables rotation}
}
\value{
The function produces a plot and does not return anything.
}
\description{
Plot the correlations between latent variables as alpha changes.
}
