% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_pathmod.R
\name{predict.pathmod}
\alias{predict.pathmod}
\title{Predict method for fitted formative-reflective models}
\usage{
\method{predict}{pathmod}(object, newdata = NULL, xmean = NULL,
  ymean = NULL, xsd = NULL, ysd = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{pathmod} generated by the function \code{pathmod()}.
If no data were supplied to the function \code{pathmod()} through the \code{data} parameter,
then the parameters}

\item{newdata}{a data frame containing the \eqn{x}-variables of the model named as they had been
named in the \code{xgroups} argument passed to the \code{pathmod()} function. If you do not supply
this argument but data have been passed to the function \code{pathmod()} through the \code{data}
parameter, then the function returs the fitted values.}

\item{xmean}{numeric vector of \eqn{x}-variable means. If no data and no \code{xmean} parameter were
supplied to the \code{pathmod()} call, then you have to assign this argument, otherwise this
parameter is not needed. If you assign it when \code{xmean} is available in \code{object},
then this value is used instead of the one present in \code{object}.}

\item{ymean}{numeric vector of \eqn{y}-variable means. If no data and no \code{ymean} parameter were
supplied to the \code{pathmod()} call, then you have to assign this argument, otherwise this
parameter is not needed. If you assign it when \code{ymean} is available in \code{object},
then this value is used instead of the one present in \code{object}.}

\item{xsd}{numeric vector of \eqn{x}-variable standard deviations. If no data and no \code{xsd} parameter were
supplied to the \code{pathmod()} call, then you have to assign this argument, otherwise this
parameter is not needed. If you assign it when \code{xsd} is available in \code{object},
then this value is used instead of the one present in \code{object}.}

\item{ysd}{numeric vector of \eqn{y}-variable means. If no data and no \code{ysd} parameter were
supplied to the \code{pathmod()} call, then you have to assign this argument, otherwise this
parameter is not needed. If you assign it when \code{ysd} is available in \code{object},
then this value is used instead of the one present in \code{object}.}

\item{...}{further arguments passed to or from other methods}
}
\value{
A list with slots:
\describe{
  \item{\code{xis}}{3-dimensional array with a matrix of \eqn{\xi}-variables for each value of \eqn{\alpha}}
  \item{\code{etas}}{3-dimensional array with a matrix of \eqn{\eta}-variables for each value of \eqn{\alpha}}
  \item{\code{xhats}}{3-dimensional array with a matrix of \eqn{x}-variable predictions for each value of \eqn{\alpha}}
  \item{\code{yhats}}{3-dimensional array with a matrix of \eqn{y}-variable predictions for each value of \eqn{\alpha}}
}
}
\description{
Once the formative-reflective model has been estimated with the \code{pathmod()} function,
you can produce predictions based on old or new data.
}
\examples{
data(russett)
xg <- list(c("gini", "farm", "rent"), c("gnpr", "labo"))
yg <- list(c("inst", "ecks", "death"))
pm <- pathmod(xg, yg, data = russett[1:40, ])
pr <- predict(pm, russett[41:47, ])
}
