% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_pathmod.R
\name{print.pathmod}
\alias{print.pathmod}
\title{Print method for fitted formative-reflective models}
\usage{
\method{print}{pathmod}(x, round = 3, ...)
}
\arguments{
\item{x}{an object of class \code{pathmod}}

\item{round}{number of decimal places for the numbers to be printed (mostly correlations and \eqn{R^2})}

\item{...}{other parameters that are passed to the standard print function used to print the matrices}
}
\value{
It only prints text, it returns nothing.
}
\description{
Print the most important results created by the function \code{pathmod}.
}
