% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrrbycov.R
\name{rrrbycov}
\alias{rrrbycov}
\title{Reduced rank regression using covariance matrices.}
\usage{
rrrbycov(Syy, Syx, Sxx, r = 1)
}
\arguments{
\item{Syy}{variance-covariance matrix of the \eqn{y} vector.}

\item{Syx}{covariance matrix of \eqn{y} and \eqn{x}.}

\item{Sxx}{variance-covariance matrix of the \eqn{x} vector.}

\item{r}{rank of reduced rank regression.}
}
\value{
A list with the following slots:
\describe{
  \item{\code{A}}{Matrix \eqn{A}}
  \item{\code{B}}{Matrix \eqn{B}}
  \item{\code{eigenvalues}}{Eigenvalues of the first \code{r} canonical correlations between \eqn{y} and \eqn{x}.}
}
}
\description{
Reduced rank regression solves the problem of estimating the matrices \eqn{A} and \eqn{B} in the linear model
\deqn{y = A B' x + \epsilon}
where \eqn{y} is a \eqn{p}-vector of response variables,
\eqn{x} is a \eqn{q}-vector of regressors and
\eqn{A} and \eqn{B} are coefficient matrices with dimensions \eqn{p x r} and \eqn{q x r}, respectively
(\eqn{1 \le r \le min(p, q)}).
The function is used internally by the package and so the passed arguments are not checked.
}
\examples{
S <- matrix(c(
  1.0, 0.2, 0.8, 0.0, 0.9, 0.9,
  0.2, 1.0, 0.0, 0.0, 0.2, 0.2,
  0.8, 0.0, 1.0, 0.2, 0.8, 0.8,
  0.0, 0.0, 0.2, 1.0, 0.1, 0.1,
  0.9, 0.2, 0.8, 0.1, 1.0, 0.8,
  0.9, 0.2, 0.8, 0.1, 0.8, 1.0),
  6,6,byrow=TRUE)
rrrbycov(S[4:6, 4:6], S[4:6, 1:3], S[1:3, 1:3], 1)
}
