% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr.R
\name{tr}
\alias{tr}
\title{Trace of a square matrix.}
\usage{
tr(M)
}
\arguments{
\item{M}{A square matrix.}
}
\value{
The trace (sum of elements on the diagonal) of \code{M}.
}
\description{
Trace of a square matrix.
}
\examples{
M <- matrix(1:9, 3, 3)
tr(M)
}
