% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pathmod.R
\name{plot.pathmod}
\alias{plot.pathmod}
\title{Plot method for fitted formative-reflective models}
\usage{
\method{plot}{pathmod}(x, ...)
}
\arguments{
\item{x}{an object of class \code{pathmod}}

\item{...}{passed to the standard plot function used inside the function to draw a scatter plot}
}
\value{
It produces a plot, it returns nothing.
}
\description{
Plot the coefficients of determination of a \code{pathmod} object for all the values in \code{alphas}.
}
